+++ common/config/bootstrap.php
@@ -8,3 +8,4 @@ Yii::setAlias('console', dirname(dirname(__DIR__)) . '/console');
 Yii::setAlias('modules', dirname(dirname(__DIR__)) . '/modules');
 Yii::setAlias('widgets', dirname(dirname(__DIR__)) . '/widgets');
 Yii::setAlias('themes', dirname(dirname(__DIR__)) . '/themes');
+Yii::setAlias('public', dirname(dirname(__DIR__)) . '/public');